+++ scripts/webpage.js
@@ -21,3 +21,9 @@ function hide_tips()
     document.getElementById("game-div").hidden = false;
     document.getElementById("tips-div").hidden = true;
 }
+
+function toggle_word_search()
+{
+    document.getElementById("word-list-div").hidden = !document.getElementById("word-list-div").hidden;
+    document.getElementById("word-search-div").hidden = !document.getElementById("word-search-div").hidden;
+}